LayoutTests/imported/w3c/web-platform-tests/webxr/render_state_vertical_fov_immersive.https-expected.txt


PASS inlineVerticalFieldOfView is set appropriately on immersively sessions - webgl
PASS inlineVerticalFieldOfView is set appropriately on immersively sessions - webgl2
